+++ build.bat
@@ -17,5 +17,5 @@ call "%OUT_DIR%\.env\Scripts\deactivate.bat"
 set "RUN_SCRIPT=%OUT_DIR%\run.bat"
 echo @echo off> "%RUN_SCRIPT%"
 echo call .env\Scripts\activate.bat>> "%RUN_SCRIPT%"
-echo python bot.py>> "%RUN_SCRIPT%"
+echo python main.py>> "%RUN_SCRIPT%"
 echo call .env\Scripts\deactivate.bat>> "%RUN_SCRIPT%"
